requirements.txt
absl-py==0.1.10
altgraph==0.15
arrow==0.12.1
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
beautifulsoup4==4.6.0
bleach==1.5.0
bs4==0.0.1
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
comtypes==1.1.4
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
cycler==0.10.0
decorator==4.2.1
docopt==0.6.2
future==0.16.0
futures==3.1.1
h5py==2.8.0rc1
html5lib==0.9999999
hyperlink==17.3.1
idna==2.6
imageio==2.3.0
incremental==17.5.0
Keras==2.1.6
lxml==4.1.1
macholib==1.9
Markdown==2.6.11
matplotlib==2.1.2
moviepy==0.2.3.4
Naked==0.1.31
networkx==2.1
nose==1.3.7
numexpr==2.6.4
numpy==1.14.0
opencv-python==3.4.0.12
pandas==0.22.0
parsel==1.3.1
pefile==2017.11.5
Pillow==3.4.2
pipreqs==0.4.9
protobuf==3.5.1
psutil==5.4.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pybrain==0.3.3
pycparser==2.18
pycryptodome==3.6.1
PyDispatcher==2.0.5
pygame==1.9.3
PyInstaller==3.3.1
PyOpenGL==3.1.0
pyOpenSSL==17.5.0
pyparsing==2.2.0
pypiwin32==223
PyQt5==5.10.1
pyqtgraph==0.10.0
python-dateutil==2.6.1
python-mode.el==6.2.3
pytz==2017.3
PyWavelets==0.5.2
pywin32==223
pywin32-ctypes==0.1.2
pywinauto==0.6.4
PyYAML==3.12
Quamash==0.6.0
queuelib==1.4.2
requests==2.18.4
requests-cache==0.4.13
robobrowser==0.5.3
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
service-identity==17.0.0
shellescape==3.4.1
sip==4.19.8
six==1.11.0
sklearn==0.0
tensorflow==1.5.0
tensorflow-tensorboard==1.5.0
tqdm==4.23.1
tushare==1.1.5
urllib3==1.22
w3lib==1.19.0
Werkzeug==0.14.1
yarg==0.1.9
zope.interface==4.4.3
